[CSL-756] Move Has* classes into Pos.Core.Class
