+++ core/Pos/Core/Block.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP                  #-}
 {-# LANGUAGE ConstraintKinds      #-}
 {-# LANGUAGE TemplateHaskell      #-}
 {-# LANGUAGE TypeFamilies         #-}
@@ -8,14 +7,6 @@ module Pos.Core.Block
        ( Blockchain (..)
        , GenericBlockHeader (..)
        , GenericBlock (..)
-
-       -- * Classes for overloaded accessors
-       , HasPrevBlock (..)
-
-       -- * Classes for headers
-       , IsHeader
-       , IsGenesisHeader
-       , IsMainHeader (..)
 
        -- * Lenses
        , gbBody
@@ -28,14 +19,10 @@ module Pos.Core.Block
        , gbhBodyProof
        ) where
 
-import           Control.Lens       (makeLenses)
+import           Control.Lens   (makeLenses)
 import           Universum
 
-import           Pos.Core.Types     (HasBlockVersion (..), HasDifficulty (..),
-                                     HasEpochIndex (..), HasHeaderHash (..),
-                                     HasSoftwareVersion (..), HeaderHash, SlotId)
-import           Pos.Crypto.Signing (PublicKey)
-import           Pos.Util.Util      (Some, applySome, liftLensSome)
+import           Pos.Core.Types (HeaderHash)
 
 ----------------------------------------------------------------------------
 -- GenericBlock
@@ -142,85 +129,6 @@ deriving instance
 makeLenses ''GenericBlockHeader
 makeLenses ''GenericBlock
 
-class HasPrevBlock s where
-    prevBlockL :: Lens' s HeaderHash
-
-instance HasPrevBlock (Some HasPrevBlock) where
-    prevBlockL = liftLensSome prevBlockL
-
 -- | Lens from 'GenericBlock' to 'BodyProof'.
 gbBodyProof :: Lens' (GenericBlock b) (BodyProof b)
 gbBodyProof = gbHeader . gbhBodyProof
-
-----------------------------------------------------------------------------
--- Classes for headers
-----------------------------------------------------------------------------
-
-#define SOME_LENS_CLASS(HAS, LENS, CL)                       \
-    instance HAS (Some CL) where LENS = liftLensSome LENS
-#define SOME_FUNC_CLASS(HAS, FUNC, CL)                       \
-    instance HAS (Some CL) where FUNC = applySome FUNC
-
--- Add (..) to export list when IsHeader or IsGenesisHeader get any methods
-
-{- | A class that lets subpackages use some fields from headers without
-depending on cardano-sl:
-
-  * 'difficultyL'
-  * 'epochIndexL'
-  * 'prevBlockL'
-  * 'headerHashG'
--}
-class (HasDifficulty header
-      ,HasEpochIndex header
-      ,HasPrevBlock header
-      ,HasHeaderHash header) =>
-      IsHeader header
-
-SOME_LENS_CLASS(HasDifficulty, difficultyL, IsHeader)
-SOME_LENS_CLASS(HasEpochIndex, epochIndexL, IsHeader)
-SOME_LENS_CLASS(HasPrevBlock,  prevBlockL,  IsHeader)
-SOME_FUNC_CLASS(HasHeaderHash, headerHash,  IsHeader)
-
-instance IsHeader (Some IsHeader)
-
--- | A class for genesis headers. Currently doesn't provide any data beyond
--- what 'IsHeader' provides.
-class IsHeader header => IsGenesisHeader header
-
-SOME_LENS_CLASS(HasDifficulty, difficultyL, IsGenesisHeader)
-SOME_LENS_CLASS(HasEpochIndex, epochIndexL, IsGenesisHeader)
-SOME_LENS_CLASS(HasPrevBlock,  prevBlockL,  IsGenesisHeader)
-SOME_FUNC_CLASS(HasHeaderHash, headerHash,  IsGenesisHeader)
-
-instance IsHeader        (Some IsGenesisHeader)
-instance IsGenesisHeader (Some IsGenesisHeader)
-
-{- | A class for main headers. In addition to 'IsHeader', provides:
-
-  * 'headerSlotL'
-  * 'headerLeaderKeyL'
-  * 'blockVersionL'
-  * 'softwareVersionL'
--}
-class (IsHeader header
-      ,HasBlockVersion header
-      ,HasSoftwareVersion header) =>
-      IsMainHeader header
-  where
-    -- | Id of the slot for which this block was generated.
-    headerSlotL :: Lens' header SlotId
-    -- | Public key of slot leader.
-    headerLeaderKeyL :: Lens' header PublicKey
-
-SOME_LENS_CLASS(HasDifficulty,      difficultyL,      IsMainHeader)
-SOME_LENS_CLASS(HasEpochIndex,      epochIndexL,      IsMainHeader)
-SOME_LENS_CLASS(HasPrevBlock,       prevBlockL,       IsMainHeader)
-SOME_FUNC_CLASS(HasHeaderHash,      headerHash,       IsMainHeader)
-SOME_LENS_CLASS(HasBlockVersion,    blockVersionL,    IsMainHeader)
-SOME_LENS_CLASS(HasSoftwareVersion, softwareVersionL, IsMainHeader)
-
-instance IsHeader     (Some IsMainHeader)
-instance IsMainHeader (Some IsMainHeader) where
-    headerSlotL = liftLensSome headerSlotL
-    headerLeaderKeyL = liftLensSome headerLeaderKeyL
