+++ core/Pos/Core/Types.hs
@@ -16,20 +16,16 @@ module Pos.Core.Types
 
         -- * ChainDifficulty
        , ChainDifficulty (..)
-       , HasDifficulty (..)
 
         -- * Version
        , ApplicationName (..)
        , BlockVersion (..)
-       , HasBlockVersion (..)
        , NumSoftwareVersion
        , SoftwareVersion (..)
-       , HasSoftwareVersion (..)
 
        -- * HeaderHash related types and functions
        , BlockHeaderStub
        , HeaderHash
-       , HasHeaderHash (..)
        , headerHashF
 
        , ProxySigLight
@@ -54,12 +50,10 @@ module Pos.Core.Types
 
         -- * Slotting
        , EpochIndex (..)
-       , HasEpochIndex (..)
        , FlatSlotId
        , LocalSlotIndex (..)
        , SlotId (..)
        , EpochOrSlot (..)
-       , HasEpochOrSlot (..)
        , slotIdF
        , epochOrSlot
 
@@ -69,7 +63,6 @@ module Pos.Core.Types
        , ScriptVersion
        ) where
 
-import           Control.Lens         (Getter, to)
 import           Crypto.Hash          (Blake2s_224)
 import           Data.Data            (Data)
 import           Data.Default         (Default (..))
@@ -82,16 +75,14 @@ import           Data.Time.Units      (Microsecond)
 import           Formatting           (Format, bprint, build, formatToString, int, ords,
                                        shown, stext, (%))
 import qualified PlutusCore.Program   as PLCore
-import           Prelude              (show)
+import qualified Prelude
 import           Serokell.AcidState   ()
 import           Serokell.Util.Base16 (formatBase16)
-import           Universum            hiding (show)
+import           Universum
 
 import           Pos.Crypto           (AbstractHash, Hash, ProxySecretKey, ProxySignature,
                                        PublicKey)
 import           Pos.Data.Attributes  (Attributes)
-import           Pos.Util.Util        (Some (..), applySome, liftLensSome)
-
 
 -- | Timestamp is a number which represents some point in time. It is
 -- used in MonadSlots and its meaning is up to implementation of this
@@ -139,13 +130,6 @@ newtype ChainDifficulty = ChainDifficulty
     { getChainDifficulty :: Word64
     } deriving (Show, Eq, Ord, Num, Enum, Real, Integral, Generic, Buildable, Typeable, NFData)
 
--- | Type class for something that has 'ChainDifficulty'.
-class HasDifficulty a where
-    difficultyL :: Lens' a ChainDifficulty
-
-instance HasDifficulty (Some HasDifficulty) where
-    difficultyL = liftLensSome difficultyL
-
 ----------------------------------------------------------------------------
 -- Version
 ----------------------------------------------------------------------------
@@ -191,18 +175,6 @@ instance Hashable BlockVersion
 instance NFData BlockVersion
 instance NFData SoftwareVersion
 
-class HasBlockVersion a where
-    blockVersionL :: Lens' a BlockVersion
-
-instance HasBlockVersion (Some HasBlockVersion) where
-    blockVersionL = liftLensSome blockVersionL
-
-class HasSoftwareVersion a where
-    softwareVersionL :: Lens' a SoftwareVersion
-
-instance HasSoftwareVersion (Some HasSoftwareVersion) where
-    softwareVersionL = liftLensSome softwareVersionL
-
 ----------------------------------------------------------------------------
 -- HeaderHash
 ----------------------------------------------------------------------------
@@ -215,15 +187,6 @@ data BlockHeaderStub
 -- | Specialized formatter for 'HeaderHash'.
 headerHashF :: Format r (HeaderHash -> r)
 headerHashF = build
-
--- | Class for something that has 'HeaderHash'.
-class HasHeaderHash a where
-    headerHash :: a -> HeaderHash
-    headerHashG :: Getter a HeaderHash
-    headerHashG = to headerHash
-
-instance HasHeaderHash (Some HasHeaderHash) where
-    headerHash = applySome headerHash
 
 ----------------------------------------------------------------------------
 -- Proxy signatures and delegation
@@ -350,13 +313,6 @@ instance Buildable EpochIndex where
 -- instance Buildable (EpochIndex,EpochIndex) where
 --     build = bprint ("epochIndices: "%pairF)
 
--- | Class for something that has 'EpochIndex'.
-class HasEpochIndex a where
-    epochIndexL :: Lens' a EpochIndex
-
-instance HasEpochIndex (Some HasEpochIndex) where
-    epochIndexL = liftLensSome epochIndexL
-
 -- | Index of slot inside a concrete epoch.
 newtype LocalSlotIndex = LocalSlotIndex
     { getSlotIndex :: Word16
@@ -403,22 +359,6 @@ instance Ord EpochOrSlot where
 
 instance Buildable EpochOrSlot where
     build = either Buildable.build Buildable.build . unEpochOrSlot
-
-class HasEpochOrSlot a where
-    _getEpochOrSlot :: a -> Either EpochIndex SlotId
-    getEpochOrSlot :: a -> EpochOrSlot
-    getEpochOrSlot = EpochOrSlot . _getEpochOrSlot
-    epochOrSlotG :: Getter a EpochOrSlot
-    epochOrSlotG = to getEpochOrSlot
-
-instance HasEpochOrSlot (Some HasEpochOrSlot) where
-    _getEpochOrSlot = applySome _getEpochOrSlot
-
-instance HasEpochOrSlot EpochIndex where
-    _getEpochOrSlot = Left
-
-instance HasEpochOrSlot SlotId where
-    _getEpochOrSlot = Right
 
 instance NFData SlotId
